Polished and added a few things
Added a different kind of a cool header, used rem instead of pt for all font sizes, made the texts a lil bigger and also added a date to each (one so far) of the projects in the the projects section. I'll later connect this website to either Firebase, Appwrite or Supabase. Not sure yet