Added some date stuff.
The projects are now sorting based on their date property and I also made the date <p> tag look better by simply putting a zero in front of the month and day of the month if the lenght of it is less than 2 (ex. instead of 1.1. 2024 it looks like 01.01. 2024 imo better)